Rename classifeds and classifiedsV2
